+++ packages/climate-ref-celery/src/climate_ref_celery/tasks.py
@@ -13,13 +13,10 @@ The main process is responsible for tracking what diagnostics have been register
 and to respond to new workers coming online.
 """
 
-from collections.abc import Callable
-
 from celery import Celery
-from loguru import logger
 
-from climate_ref_core.diagnostics import Diagnostic, ExecutionDefinition, ExecutionResult
-from climate_ref_core.logging import redirect_logs
+from climate_ref_core.diagnostics import Diagnostic
+from climate_ref_core.executor import execute_locally
 from climate_ref_core.providers import DiagnosticProvider
 
 
@@ -28,32 +25,6 @@ def generate_task_name(provider: DiagnosticProvider, diagnostic: Diagnostic) -> 
     Generate the name of the task for the given provider and diagnostic
     """
     return f"{provider.slug}.{diagnostic.slug}"
-
-
-def _diagnostic_task_factory(
-    diagnostic: Diagnostic,
-) -> Callable[
-    [ExecutionDefinition, str],
-    ExecutionResult,
-]:
-    """
-    Create a new task for the given diagnostic
-    """
-
-    def task(definition: ExecutionDefinition, log_level: str) -> ExecutionResult:
-        """
-        Task to run the diagnostic
-        """
-        logger.info(f"Running diagnostic {diagnostic.name} with definition {definition}")
-        try:
-            with redirect_logs(definition, log_level):
-                return diagnostic.run(definition)
-        except Exception:
-            logger.exception(f"Error running diagnostic {diagnostic.slug}:{definition.key}")
-            # TODO: This exception should be caught and a unsuccessful result returned.
-            raise
-
-    return task
 
 
 def register_celery_tasks(app: Celery, provider: DiagnosticProvider) -> None:
@@ -69,10 +40,14 @@ def register_celery_tasks(app: Celery, provider: DiagnosticProvider) -> None:
     provider
         The provider to register tasks for
     """
-    for metric in provider.diagnostics():
-        print(f"Registering task for diagnostic {metric.name}")
+    for diagnostic in provider.diagnostics():
+        print(f"Registering task for diagnostic {diagnostic.name}")
+
+        # The task function is the same for all diagnostics
+        # The diagnostic is included in the definition
+        # The queue is important to ensure that the task is run in the correct worker
         app.task(  # type: ignore
-            _diagnostic_task_factory(metric),
-            name=generate_task_name(provider, metric),
+            execute_locally,
+            name=generate_task_name(provider, diagnostic),
             queue=provider.slug,
         )