+++ packages/climate-ref-core/src/climate_ref_core/executor.py
@@ -2,20 +2,54 @@
 Executor interface for running diagnostics
 """
 
+import importlib
+import shutil
 from typing import TYPE_CHECKING, Any, Protocol, runtime_checkable
 
-from climate_ref_core.diagnostics import Diagnostic, ExecutionDefinition
-from climate_ref_core.providers import DiagnosticProvider
+from loguru import logger
+
+from climate_ref_core.diagnostics import ExecutionDefinition, ExecutionResult
+from climate_ref_core.exceptions import InvalidExecutorException
+from climate_ref_core.logging import redirect_logs
 
 if TYPE_CHECKING:
+    # TODO: break this import cycle and move it into the execution definition
     from climate_ref.models import Execution
 
-EXECUTION_LOG_FILENAME = "out.log"
-"""
-Filename for the execution log.
 
-This file is written via [climate_ref_core.logging.redirect_logs][].
-"""
+def execute_locally(
+    definition: ExecutionDefinition,
+    log_level: str,
+) -> ExecutionResult:
+    """
+    Run a diagnostic execution
+
+    This is the chunk of work that should be executed by an executor.
+
+    Parameters
+    ----------
+    definition
+        A description of the information needed for this execution of the diagnostic
+    log_level
+        The log level to use for the execution
+    """
+    logger.info(f"Executing {definition.execution_slug()!r}")
+
+    try:
+        if definition.output_directory.exists():
+            logger.warning(
+                f"Output directory {definition.output_directory} already exists. "
+                f"Removing the existing directory."
+            )
+            shutil.rmtree(definition.output_directory)
+        definition.output_directory.mkdir(parents=True, exist_ok=True)
+
+        with redirect_logs(definition, log_level):
+            return definition.diagnostic.run(definition=definition)
+    except Exception:
+        # If the diagnostic fails, we want to log the error and return a failure result
+        logger.exception(f"Error running {definition.execution_slug()!r}")
+        return ExecutionResult.build_from_failure(definition)
 
 
 @runtime_checkable
@@ -37,8 +71,6 @@ class Executor(Protocol):
 
     def run(
         self,
-        provider: DiagnosticProvider,
-        diagnostic: Diagnostic,
         definition: ExecutionDefinition,
         execution: "Execution | None" = None,
     ) -> None:
@@ -55,10 +87,6 @@ class Executor(Protocol):
 
         Parameters
         ----------
-        provider
-            Provider of the diagnostic
-        diagnostic
-            Diagnostic to run
         definition
             Definition of the information needed to execute a diagnostic
 
@@ -94,3 +122,46 @@ class Executor(Protocol):
         TimeoutError
             If the timeout is reached
         """
+
+
+def import_executor_cls(fqn: str) -> type[Executor]:
+    """
+    Import an executor using a fully qualified module path
+
+    Parameters
+    ----------
+    fqn
+        Full package and attribute name of the executor to import
+
+        For example: `climate_ref_example.executor` will use the `executor` attribute from the
+        `climate_ref_example` package.
+
+    Raises
+    ------
+    InvalidExecutorException
+        If the executor cannot be imported
+
+        If the executor isn't a valid `DiagnosticProvider`.
+
+    Returns
+    -------
+    :
+        Executor instance
+    """
+    module, attribute_name = fqn.rsplit(".", 1)
+
+    try:
+        imp = importlib.import_module(module)
+        executor: type[Executor] = getattr(imp, attribute_name)
+
+        # We can't really check if the executor is a subclass of Executor here
+        # Protocols can't be used with issubclass if they have non-method members
+        # We have to check this at class instantiation time
+
+        return executor
+    except ModuleNotFoundError:
+        logger.error(f"Package '{fqn}' not found")
+        raise InvalidExecutorException(fqn, f"Module '{module}' not found")
+    except AttributeError:
+        logger.error(f"Provider '{fqn}' not found")
+        raise InvalidExecutorException(fqn, f"Executor '{attribute_name}' not found in {module}")