+++ packages/climate-ref-core/src/climate_ref_core/metric_values/__init__.py
@@ -1,0 +1,16 @@
+"""
+Metric Values
+
+A metric is a single statistical evaluation contained within a diagnostic.
+A diagnostic may consist of more than one metric.
+
+Examples include bias, root mean squared error (RMSE), Earth Mover's Distance,
+phase/timing of the seasonal cycle, amplitude of the seasonal cycle, spatial or temporal correlations,
+interannual variability.
+Not all metrics are useful for all variables or should be used with every observationally constrained dataset.
+Each metric may be converted into a performance score.
+"""
+
+from .typing import ScalarMetricValue, SeriesMetricValue
+
+__all__ = ["ScalarMetricValue", "SeriesMetricValue"]