+++ packages/climate-ref-ilamb/src/climate_ref_ilamb/__init__.py
@@ -38,7 +38,7 @@ dataset_registry_manager.register(
 )
 dataset_registry_manager.register(
     "iomb",
-    base_url="https://www.ilamb.org/IOMB-Data/DATA",
+    base_url="https://www.ilamb.org/ilamb3-data/",
     package="climate_ref_ilamb.dataset_registry",
     resource="iomb.txt",
     cache_name="ilamb3",