+++ packages/climate-ref-pmp/tests/unit/test_annual_cycle.py
@@ -102,6 +102,7 @@ def test_annual_cycle_diagnostic(
     expected_reference_filename = obs4mips_data_catalog["path"].iloc[0]
 
     definition = definition_factory(
+        diagnostic=diagnostic,
         cmip6=DatasetCollection(
             pd.Series(
                 {
@@ -133,9 +134,6 @@ def test_annual_cycle_diagnostic(
     )
     # The output directory must exist
     output_dir = definition.output_directory
-    driver_script = _get_resource(
-        "pcmdi_metrics", "mean_climate/pcmdi_compute_climatologies.py", use_resources=False
-    )
     parameter_file = _get_resource(
         "climate_ref_pmp.params", "pmp_param_annualcycle_1-clims.py", use_resources=True
     )
@@ -150,8 +148,7 @@ def test_annual_cycle_diagnostic(
     # Check the first command
     cmd = result[0]
     assert cmd == [
-        "python",
-        driver_script,
+        "pcmdi_compute_climatologies.py",
         "-p",
         parameter_file,
         "--vars",
@@ -163,14 +160,12 @@ def test_annual_cycle_diagnostic(
     ]
 
     # Check the second command
-    driver_script = _get_resource("pcmdi_metrics", "mean_climate/mean_climate_driver.py", use_resources=False)
     parameter_file = _get_resource(
         "climate_ref_pmp.params", "pmp_param_annualcycle_2-metrics.py", use_resources=True
     )
     cmd = result[1]
     assert cmd == [
-        "python",
-        driver_script,
+        "mean_climate_driver.py",
         "-p",
         parameter_file,
         "--vars",