+++ packages/climate-ref/src/climate_ref/migrations/env.py
@@ -8,6 +8,11 @@ from climate_ref.models import Base, MetricValue
 from climate_ref_core.logging import capture_logging
 from climate_ref_core.pycmec.controlled_vocabulary import CV
 
+try:
+    import alembic_postgresql_enum  # noqa
+except ImportError:
+    logger.warning("alembic_postgresql_enum not installed, skipping enum migration support")
+
 # Setup logging
 capture_logging()
 logger.debug("Running alembic env")