+++ packages/climate-ref/tests/unit/test_solver.py
@@ -24,15 +24,14 @@ from climate_ref_core.diagnostics import DataRequirement, FacetFilter
 def solver(db_seeded, config) -> ExecutionSolver:
     registry = ProviderRegistry(providers=[provider])
     # Use a fixed set of providers for the test suite until we can pull from the DB
-    with db_seeded.session.begin():
-        metric_solver = ExecutionSolver.build_from_db(config, db_seeded)
+    metric_solver = ExecutionSolver.build_from_db(config, db_seeded)
     metric_solver.provider_registry = registry
 
     return metric_solver
 
 
 @pytest.fixture
-def mock_metric_execution(tmp_path, definition_factory) -> DiagnosticExecution:
+def mock_metric_execution(tmp_path, definition_factory, mock_diagnostic) -> DiagnosticExecution:
     mock_execution = mock.MagicMock(spec=DiagnosticExecution)
     mock_execution.provider = provider
     mock_execution.diagnostic = provider.diagnostics()[0]
@@ -41,7 +40,7 @@ def mock_metric_execution(tmp_path, definition_factory) -> DiagnosticExecution:
     mock_dataset_collection = mock.Mock(hash="123456", items=mock.Mock(return_value=[]))
 
     mock_execution.build_execution_definition.return_value = definition_factory(
-        execution_dataset_collection=mock_dataset_collection
+        diagnostic=mock_diagnostic, execution_dataset_collection=mock_dataset_collection
     )
     return mock_execution
 
@@ -289,8 +288,7 @@ def test_solve_metrics_default_solver(mocker, mock_metric_execution, db_seeded, 
     mock_build_solver.return_value = solver
 
     # Run with no solver specified
-    with db_seeded.session.begin():
-        solve_required_executions(db_seeded)
+    solve_required_executions(db_seeded)
 
     # Check that a result is created
     assert db_seeded.session.query(Execution).count() == 1
@@ -310,8 +308,6 @@ def test_solve_metrics_default_solver(mocker, mock_metric_execution, db_seeded, 
     # A single run would have been run
     assert mock_executor.return_value.run.call_count == 1
     mock_executor.return_value.run.assert_called_with(
-        provider=mock_metric_execution.provider,
-        diagnostic=mock_metric_execution.diagnostic,
         definition=mock_metric_execution.build_execution_definition(),
         execution=execution_result,
     )
@@ -321,8 +317,7 @@ def test_solve_metrics(mocker, db_seeded, solver, data_regression):
     mock_executor = mocker.patch.object(ExecutorConfig, "build")
     mock_build_solver = mocker.patch.object(ExecutionSolver, "build_from_db")
 
-    with db_seeded.session.begin():
-        solve_required_executions(db_seeded, dry_run=False, solver=solver)
+    solve_required_executions(db_seeded, dry_run=False, solver=solver)
 
     assert mock_build_solver.call_count == 0